SimpleFlaskAPI/requirements.txt
psycopg2==2.9.9
Flask==3.0.3
requests==2.32.3
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-instrumentation-psycopg2==0.48b0
opentelemetry-instrumentation-flask==0.48b0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-http==1.27.0
